Fix issue with [object object] persisted to index
When upserting key/val when key does not exist in a document, the result is a string "`[object object]`" getting persisted to the index. 

There are a few ways to fix this depending on what the user wants. Once case is the example in this commit that saves the object as a string (instead of the `[object object]` string. 

Alternate solution:
- line 37: ....` || ''` instead of `|| {}`